refactor: move metadata lang tag routes to hono
